% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_function.r
\name{parse_function}
\alias{parse_function}
\title{Turn JSON file of tweets into a CSV file}
\usage{
parse_function(x)
}
\arguments{
\item{x}{describe the input}
}
\value{
describe the output.
}
\description{
\code{parse_function} what does the code really do?
}
\examples{

}
